feat: introduce I/O utilities for HDF5 handling in eFEL analysis
